backend/pyproject.toml
[tool.poetry]
name = "fastapi-chat-ddd"
version = "0.1.0"
description = ""
authors = ["Daniyar <srxd1710@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
motor = "^3.4.0"
uvicorn = { extras = ["all"], version = "^0.29.0" }
pytest = "^8.2.1"
faker = "^25.2.0"
pytest-asyncio = "^0.23.7"
punq = "^0.7.0"
pydantic-settings = "^2.3.1"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]

build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_classes = "*Test"
python_files = "tests.py test_*.py"
pythonpath = "src"
testpaths = "tests"
